Change deprecated to warning tag in flavors
Changes deprecated to warning, whilst Sphinx correctly emits
class=deprecated for the <p> tag in the HTML, read the docs does not
have CSS for it. So instead use a warning tag (which it does have) for a
nice big shiny box.

This also includes a tweak for the copyright to automatically update
each year when we push updates, as it was previously 2018